Move stuff from core to backend/database. Remote unused code.
