+++ dtbase/backend/routers/sensor.py
@@ -10,9 +10,9 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from dtbase.backend.auth import authenticate_access
-from dtbase.backend.db import db_session
+from dtbase.backend.database import sensor_locations, sensors
+from dtbase.backend.database.utils import db_session
 from dtbase.backend.models import MessageResponse, SensorMeasure, SensorType, ValueType
-from dtbase.core import sensor_locations, sensors
 from dtbase.core.exc import RowMissingError
 
 router = APIRouter(
