+++ dtbase/backend/routers/user.py
@@ -6,9 +6,9 @@ from sqlalchemy.exc import IntegrityError, NoResultFound
 from sqlalchemy.orm import Session
 
 from dtbase.backend.auth import authenticate_access
-from dtbase.backend.db import db_session
+from dtbase.backend.database import users
+from dtbase.backend.database.utils import db_session
 from dtbase.backend.models import LoginCredentials, MessageResponse, UserIdentifier
-from dtbase.core import users
 
 router = APIRouter(
     prefix="/user",
